getGrenres done
